+++ packages/mobile/src/components/audio-balance-pill/AUDIOBalancePill2.tsx
