+++ packages/mobile/src/components/usdc-balance-pill/USDCBalancePill2.tsx
