+++ packages/mobile/src/components/usdc-balance-row/USDCBalanceRow2.tsx
