+++ packages/mobile/src/screens/app-drawer-screen/left-nav-drawer/LeftNavDrawer.tsx
@@ -20,8 +20,8 @@ import IconMessage from 'app/assets/images/iconMessage.svg'
 import IconSettings from 'app/assets/images/iconSettings.svg'
 import IconUpload from 'app/assets/images/iconUpload.svg'
 import IconUser from 'app/assets/images/iconUser.svg'
-import { AudioBalancePill } from 'app/components/audio-balance-pill/AudioBalancePill'
-import { USDCBalancePill } from 'app/components/usdc-balance-pill/USDCBalancePill'
+import { AudioBalancePill } from 'app/components/audio-balance-pill/AUDIOBalancePill2'
+import { USDCBalancePill } from 'app/components/usdc-balance-pill/USDCBalancePill2'
 import { useFeatureFlag, useRemoteVar } from 'app/hooks/useRemoteConfig'
 import { make, track } from 'app/services/analytics'
 import { spacing } from 'app/styles/spacing'
