+++ packages/web/src/components/nav/desktop/NavPopupMenu.tsx
@@ -22,9 +22,9 @@ import { useDispatch } from 'react-redux'
 import IconKebabHorizontal from 'assets/img/iconKebabHorizontalAlt.svg'
 import { make } from 'common/store/analytics/actions'
 import { Icon } from 'components/Icon'
-import { AudioBalancePill } from 'components/audio-balance-pill/AudioBalancePill'
+import { AudioBalancePill } from 'components/audio-balance-pill/AUDIOBalancePill2'
 import { NotificationDot } from 'components/notification-dot'
-import { USDCBalancePill } from 'components/usdc-balance-pill/USDCBalancePill'
+import { USDCBalancePill } from 'components/usdc-balance-pill/USDCBalancePill2'
 import { useIsUSDCEnabled } from 'hooks/useIsUSDCEnabled'
 import { useNavigateToPage } from 'hooks/useNavigateToPage'
 import { useFlag, useRemoteVar } from 'hooks/useRemoteConfig'
